localstorage

+++ src/components/App.js
@@ -12,14 +12,13 @@ import ShoppingCart from './Shopping_Cart'
 class App extends Component {
 
     componentDidMount() {
-        localStorage.clear()
 
         //   const userId = localStorage.getItem('userId')
         //   const userEmail = localStorage.getItem('userEmail')
 
-        //   if(userId && userEmail) {
-        //     this.props.dispatch(actions.fetchUserSuccess(userId, userEmail))
-        //   }
+          if(!userId && !userEmail) {
+              localStorage.clear()
+        }
     }
 
     render() {
